requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
Pillow==9.3.0
pygame==2.1.3.dev8
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
